requirements.txt
SpeechRecognition
pyaudio
streamlit
pymongo
python-dotenv
gtts